backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "cryptography>=44.0.2",
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "resend>=2.7.0",
    "routes>=2.5.1",
    "ruff>=0.11.2",
    "sqlalchemy>=2.0.40",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.0",
]
